+++ SharpMonEngine.Battle.Core/Model/SpeciesBattleInstance.cs
@@ -11,15 +11,15 @@ namespace SharpMonEngine.Battle.Core.Model
         public MonType Type2 => _speciesInstance.Type2;
         public int Level => _speciesInstance.Level;
         public int CurrentHp { get; set; }
-        public double CurrentAtk => GetCurrentStatValue(AtkLevel, _speciesInstance.Atk);
+        public float CurrentAtk => GetCurrentStatValue(AtkLevel, _speciesInstance.Atk);
         public int AtkLevel { get; set; }
-        public double CurrentDef => GetCurrentStatValue(DefLevel, _speciesInstance.Def);
+        public float CurrentDef => GetCurrentStatValue(DefLevel, _speciesInstance.Def);
         public int DefLevel { get; set; }
-        public double CurrentSpAtk => GetCurrentStatValue(SpAtkLevel, _speciesInstance.SpAtk);
+        public float CurrentSpAtk => GetCurrentStatValue(SpAtkLevel, _speciesInstance.SpAtk);
         public int SpAtkLevel { get; set; }
-        public double CurrentSpDef => GetCurrentStatValue(SpDefLevel, _speciesInstance.SpDef);
+        public float CurrentSpDef => GetCurrentStatValue(SpDefLevel, _speciesInstance.SpDef);
         public int SpDefLevel { get; set; }
-        public double CurrentSpeed => GetCurrentStatValue(SpeedLevel, _speciesInstance.Speed);
+        public float CurrentSpeed => GetCurrentStatValue(SpeedLevel, _speciesInstance.Speed);
         public int SpeedLevel { get; set; }
         public int EvasionLevel { get; set; }
         public byte Status { get; set; }
@@ -29,36 +29,36 @@ namespace SharpMonEngine.Battle.Core.Model
             _speciesInstance = speciesInstance;
         }
 
-        private double GetCurrentStatValue(int level, int baseValue)
+        private float GetCurrentStatValue(int level, int baseValue)
         {
             switch (level)
             {
                 case -6:
-                    return baseValue * 2.0 / 8;
+                    return baseValue * 2.0f/ 8;
                 case -5:
-                    return baseValue * 2.0 / 7;
+                    return baseValue * 2.0f/ 7;
                 case -4:
-                    return baseValue * 2.0 / 6;
+                    return baseValue * 2.0f/ 6;
                 case -3:
-                    return baseValue * 2.0 / 5;
+                    return baseValue * 2.0f/ 5;
                 case -2:
-                    return baseValue * 2.0 / 4;
+                    return baseValue * 2.0f/ 4;
                 case -1:
-                    return baseValue * 2.0 / 3;
+                    return baseValue * 2.0f/ 3;
                 case 0:
-                    return baseValue * 2.0 / 2;
+                    return baseValue * 2.0f/ 2;
                 case 1:
-                    return baseValue * 3.0 / 2;
+                    return baseValue * 3.0f/ 2;
                 case 2:
-                    return baseValue * 4.0 / 2;
+                    return baseValue * 4.0f / 2;
                 case 3:
-                    return baseValue * 5.0 / 2;
+                    return baseValue * 5.0f / 2;
                 case 4:
-                    return baseValue * 6.0 / 2;
+                    return baseValue * 6.0f / 2;
                 case 5:
-                    return baseValue * 7.0 / 2;
+                    return baseValue * 7.0f / 2;
                 case 6:
-                    return baseValue * 8.0 / 2;
+                    return baseValue * 8.0f / 2;
                 default:
                     return baseValue;
             }
